Updated health bar. Fixed a bug where spell lines were rendered behind the background.

+++ SpellSlider/Assets/Scripts/AdventureView/Wizard.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 
 public class Wizard : MonoBehaviour {
 	
-	public int wizardHealth = 100;
+	public float wizardHealth = 100;
 	public bool alive = true;
 
 	// Use this for initialization
